Set router peer comment format and expose config name after QR

+++ bot/models.py
@@ -2,7 +2,7 @@
 SQLAlchemy database models.
 """
 from datetime import datetime
-from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Boolean, Text, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, Float, String, DateTime, Boolean, Text, ForeignKey
 from sqlalchemy.orm import relationship
 from database import Base
 
@@ -62,7 +62,7 @@ class Plan(Base):
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, nullable=False)
     duration_days = Column(Integer, nullable=False)
-    traffic_gb = Column(Integer, nullable=False)
+    traffic_gb = Column(Float, nullable=False)
     price = Column(Integer, nullable=False)
     description = Column(Text, nullable=True)
     is_active = Column(Boolean, default=True)
